+++ src/constellation/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["__version__"]
 
-__version__ = "7.1.1"
+__version__ = "7.1.2"